Cleanup live migration and add task queue module.
